fix: split response-stream renderer tests for backend-size limit
Move response-stream renderer tests out of run_config_test.go so pkg/cli/run stays under the 1000-line backend-size cap required by CI.
